Small refactor of components
