+++ client/src/components/TodoPage.tsx
@@ -1,0 +1,35 @@
+import PaginationBar from "./PaginationBar";
+import TodoTable from "./TodoTable";
+
+//import { useDispatch, useSelector } from "react-redux";
+//import { AppDispatch, RootState } from "../redux/store";
+import { ToDo } from "../types/todoTypes";
+import { useState } from "react";
+// import { fetchToDos, removeTodo, toggleTodo } from "../redux/slices/todoSlice";
+import TodoModal from "./TodoModal";
+
+
+export default function TodoPage() {
+    //const dispatch = useDispatch<AppDispatch>();
+    const [isModalOpen, setModalOpen] = useState(false);
+    const [currentTodo, setCurrentTodo] = useState<ToDo | null>(null);
+
+    const handleEdit = (todo: ToDo) => {
+        setCurrentTodo(todo);
+        setModalOpen(true);
+    }
+
+    const closeModal = () => {
+        setModalOpen(false);
+        setCurrentTodo(null);
+    }
+
+    
+    return (
+        <>
+            <TodoTable handleEdit={handleEdit} />
+            <TodoModal isOpen={isModalOpen} onClose={closeModal} todo={currentTodo} isEditing={currentTodo !== null} />
+            <PaginationBar />
+        </>
+    )
+}
